Update Readme; Refactor test suite
