+++ README.md
@@ -2,28 +2,27 @@
 
 ## Bacteria Tracker
 
-In this exercise, your task to build a workflow for testing people for a bacterial infection and then making sure that a clinic has a report of people who have it.
+In this exercise, your task is to build a workflow for testing people for a bacterial infection, and then making sure that a clinic has a report of people who have it.
 
 ## Setup
 
 1. Open your terminal and `cd` to your workspace directory.
 1. Clone this project.
-1. Once the project is cloned, `cd clinic` and then run `npm install`.
-1. Once the installations are complete, run the `npm run test` command. You will see all of the tests for your code - which ones pass, and which ones fail. As soon as you make changes to your code, the tests will run automatically, so keep this terminal window open while you are working on this project. When you need to stop working on it, use the `Ctrl+C` keyboard shortcut to stop the tests from running.
+1. Once the project is cloned, cd into it and then run `npm install`.
+1. Once the installations are complete, run the `npm run test` command. You will see all of the tests for your code - which ones pass, and which ones fail. All the tests will fail at this point, because you haven't written the code to make them pass. Keep this terminal window open while you are working on this project. Come back to it to run the tests again after adding each feature described below.
 
-
-> **Note:** All of the code for this project will be created the `src` directory of the project. Open a new terminal session, and `cd` to the project directory again. Then you can `cd src` to change to that directory in your terminal. To check what's already in the `src` directory, use the `ls` command to see its contents.
+> **Note:** All of the code for this project will be created the `src` directory of the project. Open a new terminal session, and `cd` to the project directory again. Then you can `cd src` to change to that directory in your terminal and run your appplication server. To check what's already in the `src` directory, use the `ls` command to see its contents.
 
 ## Testing the Patients
 
 1. Create a `scripts/TestFacility.js` module.
-1. Define a variable in the module to have the hold of the primary key for each patient _(i.e. the `id` property)_. It should have an initial value of 1.
+1. Define a variable in the module to hold a number that you can use when setting the primary key for each person you create _(i.e. the `id` property)_. It should have an initial value of 1. Use this when setting the id property on the person. You will need to increase this value by one after each new person is created.
 1. Define and export a function named `testPerson`.
 1. The `testPerson` function must accept the following values as input _(i.e. it needs parameters)_, in the following order.
     1. First name of the person being tested (e.g. "Kelly", "Peter")
     1. Age of the person (e.g. 31, 65)
     1. Person's temperature (e.g. 98, 103)
-1. The `testPerson` function must return an object with the following properties on it. The `id` value should be incremented by 1 each time a person is tested.
+1. The `testPerson` function must return an object with the following properties on it. The `id` value you defined earlier should be incremented by 1 each time a person is tested.
     1. `firstName` whose value comes from the parameter
     1. `age` whose value comes from the parameter
     1. `temperature` whose value comes from the parameter
@@ -42,9 +41,9 @@ Once you have it working, test 5 people in the `main.js` module.
 ## Clinical Consultation
 
 1. Define a `scripts/Clinic.js` module.
-1. Define a variable in the module that will store the people objects after they have been diagnosed in the clinic. It's initial value should be an empty array.
+1. Define a variable in the module that will store the people objects after they have been diagnosed in the clinic. Its initial value should be an empty array.
 1. Define and export a function named `usePatients` which should return the array of diagnosed patients.
-1. Define and export a function named `diagnose` that is responsible providing a diagnosis for a tested person.
+1. Define and export a function named `diagnose` that is responsible for providing a diagnosis for a tested person.
 1. The function must accept the following values as input _(i.e. it needs parameters)_, in the following order.
     1. An object representing a person who was tested with the `testPerson` function.
     1. A number specifying how many days the person has been exhibiting symptoms.
@@ -57,7 +56,7 @@ Once you have it working, test 5 people in the `main.js` module.
 
 #### Checking Your Work
 
-In the `main.js` module, invoke the `diagnose` function for each of the 5 people who were test. Ensure you provide the required values as arguments. Store the object that gets returned into a variable, and then use `console.log()` to view the objects and make sure it has the right properties on each.
+In the `main.js` module, invoke the `diagnose` function for each of the 5 people who were tested. Ensure you provide the required values as arguments. Store the object that gets returned into a variable, and then use `console.log()` to view the objects and make sure it has the right properties on each.
 
 To check your work, make sure that at least one of the people is infected by providing a temperature that is too high, and a number symptomatic days being 4 or greater.
 
@@ -85,7 +84,7 @@ Your next task is to create HTML representations of the people who have been tes
 1. Define and export a `PatientList` function.
 1. The `PatientList` function must import that array of patients from the from the `Clinic.js` module.
 1. The `PatientList` function must convert each object in the array to an HTML representation string.
-Use the following template to generate the representations. Recall the `${}` syntax for interpolating JavaScript variables into string templates.
+The resulting HTML should look like the following example. Recall the `${}` syntax for interpolating JavaScript variables into string templates.
     ```html
     <section class="patient" id="patient--1">
         <h2 class="patient__name">Doug</h2>
@@ -99,14 +98,13 @@ Use the following template to generate the representations. Recall the `${}` syn
         </div>
     </section>
     ```
-1. The function must put all of the HTML representations into a single string.
+1. The function must put all of the HTML representations into a single string. Recall that there are ways to turn an array's values into a single string.
 1. The function then must return that single string that has all of the patient HTML representation in it.
-1. In the `main.js` module, invoke the `PatientList` component function. Take the string of HTML representations that it returns and update the inner HTML of article element you created in `index.html`. You need to remember how to get a reference to DOM element in JavaScript and then assign it some HTML.
+1. In the `main.js` module, invoke the `PatientList` component function. Take the string of HTML representations that it returns and update the inner HTML of the article element you created in `index.html`. You need to remember how to get a reference to DOM element in JavaScript and then assign it some HTML.
 
 #### Checking Your Work
 
-
-Make sure your web server is running, and then visit http://localhost:8080 and you should see your list of diagnosed patient. It should look similiar to this.
+Make sure your web server is running, and then visit http://localhost:<yourPort#> and you should see your list of diagnosed patients. It should look similiar to this.
 
 ![](./patients.png)
 
@@ -115,4 +113,3 @@ Then look at your terminal window that is running the tests and make sure that t
 * `Patients are rendered to DOM`
 
 **THEN PUSH YOUR CODE TO GITHUB**
-
